Tests fixes for geofencing and location
