+++ lib/agent/providers/geo/index.js
@@ -29,7 +29,7 @@ exports.fetch_location = function(cb) {
       log_error(err, "google");
       return strategies.geoip(geoip_cb);
     }
-    logger.info("GOOGLE: " + JSON.stringify(res))
+
     return cb(null, res);
   }
